// StudentHousing/StudentHousing.ClientApp/src/models/formInterfaces/newAdvertisementFormData.ts
export interface NewAdvertisementFormData {
  categoryName: string;
  region: string;
  postalCode: string;
  city: string;
  district: string;
  streetName: string;
  streetNumber: string;
  unitNumber: string;
  numberOfRooms: string;
  size: string;
  furnished: string;
  parking: string;
  description: string;
  monthlyPrice: string;
}